requirements.txt
apache-airflow-providers-apache-spark==4.1.1
pyspark==3.2.1
